Create bc layer
