+++ docs/resource_factories.md
@@ -128,7 +128,7 @@ declare(strict_types=1);
 
 namespace App\Factory;
 
-use App\Entity\Book;use Sylius\Resource\Doctrine\Peristence\RepositoryInterface;use Sylius\Resource\Factory\FactoryInterface;
+use App\Entity\Book;use Sylius\Resource\Doctrine\Persistence\RepositoryInterface;use Sylius\Resource\Factory\FactoryInterface;
 
 final class BookFactory implements FactoryInterface
 {
