+++ src/Component/Repository/Exception/ExistingResourceException.php
@@ -1,0 +1,22 @@
+<?php
+
+/*
+ * This file is part of the Sylius package.
+ *
+ * (c) Sylius Sp. z o.o.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Sylius\Component\Resource\Repository\Exception;
+
+class_exists(\Sylius\Resource\Doctrine\Persistence\Exception\ExistingResourceException::class);
+
+if (false) {
+    class ExistingResourceException extends \Sylius\Resource\Doctrine\Persistence\Exception\ExistingResourceException
+    {
+    }
+}
