+++ src/Component/tests/spec/Symfony/Request/State/ProviderSpec.php
@@ -16,10 +16,10 @@ namespace spec\Sylius\Resource\Symfony\Request\State;
 use Pagerfanta\Pagerfanta;
 use PhpSpec\ObjectBehavior;
 use Psr\Container\ContainerInterface;
-use Sylius\Resource\Doctrine\Peristence\RepositoryInterface;
 use Sylius\Component\Resource\Tests\Dummy\RepositoryWithCallables;
 use Sylius\Resource\Context\Context;
 use Sylius\Resource\Context\Option\RequestOption;
+use Sylius\Resource\Doctrine\Persistence\RepositoryInterface;
 use Sylius\Resource\Metadata\Index;
 use Sylius\Resource\Metadata\Operation;
 use Sylius\Resource\Symfony\ExpressionLanguage\ArgumentParserInterface;
