+++ tests/Application/src/Repository/ComicBookRepository.php
@@ -17,7 +17,7 @@ use App\Entity\ComicBook;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
 use Doctrine\Persistence\ManagerRegistry;
 use Sylius\Bundle\ResourceBundle\Doctrine\ORM\ResourceRepositoryTrait;
-use Sylius\Resource\Doctrine\Peristence\RepositoryInterface;
+use Sylius\Resource\Doctrine\Persistence\RepositoryInterface;
 
 final class ComicBookRepository extends ServiceEntityRepository implements RepositoryInterface
 {
